refactor(메뉴): 메뉴, 메뉴그룹 도메인 분리
