+++ src/main/java/kitchenpos/eatinorders/domain/OrderLineItem.java
@@ -1,6 +1,6 @@
 package kitchenpos.eatinorders.domain;
 
-import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.menus.tobe.domain.menu.Menu;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
