+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,8 +1,7 @@
 package kitchenpos.menus.application;
 
-import kitchenpos.menus.dto.MenuGroupResponse;
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.domain.menu.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
