+++ src/main/java/kitchenpos/menus/dto/MenuGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.dto;
 
-import kitchenpos.menus.tobe.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.menu.MenuGroup;
 
 import java.util.UUID;
