+++ src/main/java/kitchenpos/menus/dto/MenuResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.dto;
 
-import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.menus.tobe.domain.menu.Menu;
 
 import java.math.BigDecimal;
 import java.util.Collections;
