+++ src/main/java/kitchenpos/menus/infra/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.infra;
 
+import kitchenpos.menus.tobe.domain.menu.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
