+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuDisplayedName.java
@@ -1,24 +1,33 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import kitchenpos.menus.exception.MenuDisplayedNameException;
 import kitchenpos.menus.exception.MenuErrorCode;
+import kitchenpos.common.domain.ProfanityPolicy;
 
 import javax.persistence.Embeddable;
 import java.util.Objects;
 
 @Embeddable
-public class MenuGroupDisplayedName {
+public class MenuDisplayedName {
     private String name;
 
-    protected MenuGroupDisplayedName() {
+    protected MenuDisplayedName() {
 
     }
 
-    public MenuGroupDisplayedName(String name) {
-        if (isNullAndEmpty(name)) {
+    public MenuDisplayedName(String name, ProfanityPolicy profanityPolicy) {
+        validate(name, profanityPolicy);
+        this.name = name;
+    }
+
+    private void validate(String text, ProfanityPolicy profanityPolicy) {
+        if (isNullAndEmpty(text)) {
             throw new MenuDisplayedNameException(MenuErrorCode.NAME_IS_NULL_OR_EMPTY);
         }
-        this.name = name;
+
+        if (profanityPolicy.containsProfanity(text)) {
+            throw new MenuDisplayedNameException(MenuErrorCode.NAME_HAS_PROFANITY);
+        }
     }
 
     private boolean isNullAndEmpty(String name) {
@@ -33,7 +42,7 @@ public class MenuGroupDisplayedName {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        MenuGroupDisplayedName that = (MenuGroupDisplayedName) o;
+        MenuDisplayedName that = (MenuDisplayedName) o;
         return Objects.equals(name, that.name);
     }
