+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuProductQuantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import kitchenpos.menus.exception.MenuErrorCode;
 import kitchenpos.menus.exception.MenuProductQuantityException;
