+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
+
+import kitchenpos.menus.tobe.domain.menu.MenuGroup;
 
 import java.util.List;
 import java.util.Optional;
