+++ src/test/java/kitchenpos/menus/tobe/domain/MenuGroupDisplayedNameTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.menus.tobe.domain;
 
 import kitchenpos.menus.exception.MenuErrorCode;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupDisplayedName;
 import kitchenpos.products.exception.ProductDisplayedNameException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
