+++ src/test/java/kitchenpos/menus/tobe/domain/MenuTest.java
@@ -3,6 +3,9 @@ package kitchenpos.menus.tobe.domain;
 import kitchenpos.common.domain.Price;
 import kitchenpos.menus.application.FakeMenuProfanityPolicy;
 import kitchenpos.common.exception.PriceException;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuProduct;
+import kitchenpos.menus.tobe.domain.menu.MenuProducts;
 import kitchenpos.products.tobe.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
